[SW-48] feat: useToggleMark 커스텀 훅 구현
